requirements.txt
# LangChain Core
langchain
langchain-core
langchain-community

# OpenAI Integration
# langchain-openai
# openai

# Anthropic Integration for claude
# langchain-anthropic

# Google Gemini Integration
langchain-google-genai
google-generativeai

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn

# for embedding models
sentence-transformers

# for web application
streamlit

# for strutured output
pydantic

# for visualizing chains
grandalf


# for password protected PDF files
cryptography